# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/courses/{id}": {
      "get": {
        "operationId": "CourseController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course details."
          },
          "404": {
            "description": "Course not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a course by ID",
        "tags": [
          "Courses"
        ]
      },
      "put": {
        "operationId": "CourseController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Course updated."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Course not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a course (Admin only)",
        "tags": [
          "Courses"
        ]
      },
      "delete": {
        "operationId": "CourseController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Course ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Course deleted."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Course not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a course (Admin only)",
        "tags": [
          "Courses"
        ]
      }
    },
...
    "/assessments/content/{contentId}": {
      "post": {
        "operationId": "AssessmentController_create",
        "parameters": [
          {
            "name": "contentId",
            "required": true,
            "in": "path",
            "description": "Course Content ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssessmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Assessment created successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new assessment for a course content",
        "tags": [
          "Assessments"
        ]
      }
    },
    "/assessments/{id}": {
      "put": {
        "operationId": "AssessmentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Assessment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAssessmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update assessment basics",
        "tags": [
          "Assessments"
        ]
      },
      "get": {
        "operationId": "AssessmentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Assessment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get assessment details including quiz or code challenge",
        "tags": [
          "Assessments"
        ]
      }
    },
    "/assessments/{id}/quiz": {
      "post": {
        "operationId": "AssessmentController_createQuiz",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Assessment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a quiz to an assessment",
        "tags": [
          "Assessments"
        ]
      }
    },
    "/assessments/quiz/{quizId}/questions": {
      "post": {
        "operationId": "AssessmentController_addQuestion",
        "parameters": [
          {
            "name": "quizId",
            "required": true,
            "in": "path",
            "description": "Quiz ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizQuestionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a question to a quiz",
        "tags": [
          "Assessments"
        ]
      }
    },
    "/assessments/{id}/code-challenge": {
      "post": {
        "operationId": "AssessmentController_createCodeChallenge",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Assessment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCodeChallengeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a code challenge to an assessment",
        "tags": [
          "Assessments"
        ]
      }
    },
    "/assessments/sessions": {
      "post": {
        "operationId": "AssessmentController_createSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateQuizSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quiz session created."
          },
          "404": {
            "description": "Quiz or cohort not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new group quiz session",
        "tags": [
          "Assessments"
        ]
      }
    },
    "/assessments/sessions/{sessionId}/status": {
      "patch": {
        "operationId": "AssessmentController_updateStatus",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Quiz Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "waiting",
                      "active",
                      "completed"
                    ],
                    "description": "New session status"
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session status updated."
          },
          "404": {
            "description": "Session not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update status of a quiz session",
        "tags": [
          "Assessments"
        ]
      }
    },
    "/assessments/sessions/{sessionId}/submit": {
      "post": {
        "operationId": "AssessmentController_submitAnswer",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Quiz Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizAnswerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Answer submitted."
          },
          "404": {
            "description": "Session not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit an answer to an active quiz session",
        "tags": [
          "Assessments"
        ]
      }
    },
    "/assessments/sessions/{sessionId}/leaderboard": {
      "get": {
        "operationId": "AssessmentController_getLeaderboard",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "description": "Quiz Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard data."
          },
          "404": {
            "description": "Session not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current leaderboard for a quiz session",
        "tags": [
          "Assessments"
        ]
      }
    },
    "/assessments/code-challenge/{challengeId}/submit": {
      "post": {
        "operationId": "AssessmentController_submitCode",
        "parameters": [
          {
            "name": "challengeId",
            "required": true,
            "in": "path",
            "description": "Code Challenge ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCodeChallengeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Code submitted and evaluated."
          },
          "404": {
            "description": "Challenge not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit code for a challenge and evaluate",
        "tags": [
          "Assessments"
        ]
      }
    },
    "/assessments/code-challenge/{challengeId}/test": {
      "post": {
        "operationId": "AssessmentController_testCode",
        "parameters": [
          {
            "name": "challengeId",
            "required": true,
            "in": "path",
            "description": "Code Challenge ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCodeChallengeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Test results returned."
          },
          "404": {
            "description": "Challenge not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Test code challenge without creating a submission",
        "tags": [
          "Assessments"
        ]
      }
    },
...
  },
  "info": {
    "title": "LMS API",
    "description": "## Learning Management System REST API\n\nThis API powers the LMS platform, supporting:\n- **Authentication** – signup, login, password reset\n- **Users** – CRUD for all user roles (SuperAdmin, Admin, Tutor, Learner)\n- **Courses** – course management, tutor assignment, learner enrollment, content tree\n- **Assessments** – quizzes, code challenges, group Kahoot-style sessions\n- **AI** – audio transcription, quiz/flashcard generation, performance analysis\n\n### Authentication\nAll protected endpoints require a **Bearer JWT token**.  \nObtain a token via `POST /auth/login` and pass it as:\n```\nAuthorization: Bearer <token>\n```\n\n### Roles\n| Role | Description |\n|------|-------------|\n| `superadmin` | Full access |\n| `admin` | Manage courses, users, assessments |\n| `tutor` | Manage assigned course content |\n| `learner` | View content, submit assessments |\n",
    "version": "1.0",
    "contact": {
      "name": "LMS Team",
      "url": "",
      "email": "support@lms.example.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter your JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Advanced TypeScript",
            "description": "Course title"
          },
          "description": {
            "type": "string",
            "example": "Deep dive into TypeScript",
            "description": "Course description"
          },
          "difficultyLevel": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "description": "Difficulty level"
          }
        }
      },
...
      "CreateAssessmentDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "quiz",
              "assignment",
              "group_assignment",
              "code_challenge",
              "kahoot_quiz"
            ],
            "description": "Type of assessment"
          },
          "title": {
            "type": "string",
            "example": "Final Quiz",
            "description": "Title of the assessment"
          },
          "instructions": {
            "type": "string",
            "example": "Please answer all questions carefully.",
            "description": "Instructions for the assessment"
          }
        },
        "required": [
          "type",
          "title"
        ]
      },
      "UpdateAssessmentDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "quiz",
              "assignment",
              "group_assignment",
              "code_challenge",
              "kahoot_quiz"
            ],
            "description": "Type of assessment"
          },
          "title": {
            "type": "string",
            "example": "Updated Quiz Title",
            "description": "Title of the assessment"
          },
          "instructions": {
            "type": "string",
            "example": "Updated instructions.",
            "description": "Instructions for the assessment"
          }
        }
      },
      "CreateQuizDto": {
        "type": "object",
        "properties": {
          "timeAllocated": {
            "type": "number",
            "example": 30,
            "description": "Time allocated in minutes"
          },
          "passMark": {
            "type": "number",
            "example": 70,
            "description": "Pass mark percentage"
          },
          "isGroup": {
            "type": "boolean",
            "example": false,
            "description": "Is it a group quiz?"
          },
          "groupPin": {
            "type": "string",
            "example": "123456",
            "description": "PIN for group sessions"
          }
        },
        "required": [
          "timeAllocated",
          "passMark"
        ]
      },
      "CreateQuizQuestionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "single_option",
              "multiple_option",
              "true_false",
              "image_matching",
              "fill_in_the_blank",
              "short_answer"
            ],
            "description": "Type of question"
          },
          "question": {
            "type": "string",
            "example": "What is 2+2?",
            "description": "The question text"
          },
          "options": {
            "type": "object",
            "example": [
              "1",
              "2",
              "3",
              "4"
            ],
            "description": "Options for multiple choice"
          },
          "correctAnswer": {
            "type": "object",
            "example": "4",
            "description": "The correct answer"
          },
          "marks": {
            "type": "number",
            "example": 1,
            "description": "Marks for this question"
          },
          "timeLimit": {
            "type": "number",
            "example": 30,
            "description": "Time limit for this question in seconds"
          }
        },
        "required": [
          "type",
          "question",
          "correctAnswer"
        ]
      },
      "CreateCodeChallengeDto": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "example": "typescript",
            "description": "Programming language"
          },
          "problemStatement": {
            "type": "string",
            "example": "Write a function that returns the sum of two numbers.",
            "description": "Problem statement"
          },
          "boilerplateCode": {
            "type": "string",
            "example": "function sum(a, b) {\n  return a + b;\n}",
            "description": "Boilerplate code"
          },
          "testCases": {
            "example": [
              {
                "input": "1, 2",
                "expectedOutput": "3",
                "isHidden": false
              }
            ],
            "description": "Test cases",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "language",
          "problemStatement",
          "testCases"
        ]
      },
      "CreateQuizSessionDto": {
        "type": "object",
        "properties": {
          "quizId": {
            "type": "string",
            "example": "uuid-of-quiz",
            "description": "ID of the quiz"
          },
          "cohortId": {
            "type": "string",
            "example": "uuid-of-cohort",
            "description": "ID of the cohort"
          }
        },
        "required": [
          "quizId",
          "cohortId"
        ]
      },
      "SubmitQuizAnswerDto": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string",
            "example": "uuid-of-question",
            "description": "ID of the question"
          },
          "answer": {
            "type": "object",
            "example": "Option A",
            "description": "The chosen answer"
          },
          "responseTimeMs": {
            "type": "number",
            "example": 1500,
            "description": "Response time in milliseconds"
          }
        },
        "required": [
          "questionId",
          "answer",
          "responseTimeMs"
        ]
      },
      "SubmitCodeChallengeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "function sum(a, b) {\n  return a + b;\n}",
            "description": "The submitted code"
          }
        },
        "required": [
          "code"
        ]
      },
...
    }
  }
}